perf(users): optimize user stats and activity series
- Use countByUserId/Status for reservation stats
- Add user-scoped payment/refund count and sum queries
- Compute monthly series via DB aggregation (no in-memory scans)
- Return BigDecimal for spending; update DTO accordingly
- Remove global queries; focus on current user only

+++ src/main/java/com/adamo/vrspfab/users/UserActivitySeriesDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -14,7 +15,7 @@ import java.util.List;
 public class UserActivitySeriesDto {
     private List<String> months; // e.g., ["Apr", "May", ...]
     private List<Long> monthlyReservationCounts;
-    private List<Double> monthlySpending; // completed payment amounts per month
+    private List<BigDecimal> monthlySpending; // completed payment amounts per month
 }
